% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim.R
\docType{data}
\name{clim}
\alias{clim}
\title{Climate Data from Santa Barbara}
\format{A data frame with 27274 rows and 8 colums
\itemize{
\item date day/month/year
\item tmin minimum daily temperature (C)
\item tmax maximum daily temperature (C)
\item rain daily rainfall (mm)
\item year
\item month
\item day
\item wy water year
}}
\source{
\url{http://sbs.lternet.edu/data/}
}
\usage{
clim
}
\description{
Data from SB-LTER meterology station
}
\author{
Naomi
}
\keyword{datasets}
