% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spring_summary.R
\name{spring_summary}
\alias{spring_summary}
\title{Summary information about spring climate}
\usage{
spring_summary(clim_data, spring_months = c(4:6), springout = FALSE)
}
\arguments{
\item{clim_data}{data frame with columns tmax, tmin (C)
rain (precip in mm), year, month (integer), day}

\item{months}{(as integer) to include in spring; default 4,5,6}
}
\value{
returns a list containing, mean spring temperature (mean_springT, (C))
year with lowest spring temperature (coldest_spring (year))
mean spring precipitation (mean_springP (mm))
spring (as year) with highest precip (wettest_spring (year))
}
\description{
computes summary information about spring temperature and precipitation
}
