% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surge_to_damage.R
\name{surge_to_damage}
\alias{surge_to_damage}
\title{surge_to_damage}
\usage{
surge_to_damage(surge, surge.min, base, K)
}
\arguments{
\item{surge}{(m)}

\item{surge.min}{(m) minimum surge that causes damage}

\item{base}{($) costs associated with any surge above the minimum}

\item{K}{($/m) slope of the surge/damage linear relationship}
}
\value{
damage in $
}
\description{
Function to compute the costs of damages as a function of surge 
height
}
